services/enricher/requirements.txt
nltk
python-dateutil
kafka-python

